+++ R/zzz.R
@@ -10,3 +10,8 @@
 	# 		'Not using Cholesky decomposition (safe).')
 	# packageStartupMessage(paste0('Package bayessource loaded. ', s))
 }
+
+# Detach the DLL when unloading the library
+.onUnload <- function(libpath) {
+   library.dynam.unload('bayessource', libpath)
+}